Fix Diary view refresh on first select
When youi switch to Diary view no date range is known by the
charts because they have just been created. Since we have a
lazy create we need to resignal the date range after they
are created.

This was already in MainWindow with a call to GcCalendar::refresh
but a recent update optimised that out.