Fix for 'stuck in full screen' issue
When GC is restarted after being closed full screen, the saved geometry
is larger than the actual screen size - resulting in the title bar being
drawn off screen & not easily reachable.

This patch checks whether the saved geometry is larger than the current
screen size, and resizes to the screen size instead of the saved values.