Don't start in full screen]
If you close or crash in fullscreen mode then it is
restored on startup but not in fullscreen mode. So
it is impossible to shrink -- this is particularly
relevant on Windows 8.

We now restrict the initial size to always be 60px
smaller in x and y than the actual screen size.