Added Upgrade processing
We now maintain a build id which correlates to a release
version (or development version if we want).

Most importantly, when we increment the build id we get to
run some processing when the cyclist is opened to clean/prep
for this release.

I have added V3 RC3 as the first build id and it will
remove *.cpi and *.bak files as well as removing the
clucene index and metricDBv3 file.

Fixes #584.