Attempt to fix intermittent segfault at end of workout
There is an intermittent segfault in ANTLogger::logRawAntMessage() at
the end of a workout. When this occurs it seems that the message data
can not be read.

This commit just adds a 500ms delay before closing the ANT worker thread,
which should allow any queued messages to be written to the log file
before the thread terminates.