Fix delete ride SEGV
Ridesummarywindow was using rideitem to get zones to use when
summarising for a date range which was bad because a) the ride
might not be for the date range selected and b) it might have
been deleted.

This fixes that and means we get a summary that is more accurate
and no crashing when rides are deleted.