Add KK roller wheel size option
Adds another option the the selection of wheel sizes: the 2.125" roller
on Kurt Kinetic trainers. Should be 169.56mm, but rounded up to the
nearest integer as 170mm.
The reason for adding this is to support users of Kurt's InRide sensor,
or people putting an ANT+ sensor in the same location.