Handle SSL errors when using Google Maps API
The default.css location now redirects to an HTTPS URL, causing an
SSL error to be thrown in some builds. This results in the style
sheet not being loaded, and the map height becoming 0 pixels and
therefore not rendering correctly.

This patch displays the state of SSL support, and if any errors are
thrown, the error messages and the peer certificate.

It will then ignore the error, allowing the Google Maps page to be
displayed correctly.