Fix CP chart defaults
The CP chart model defaults are set before the spin box step
min and max values are set, which means that the value is always
reset to 100.

Since the vast majority of users will not be aware of the settings
and/or will never think of setting them to their own 'preferences'
it is vital the default is applied correctly.