Fix events shown on LTM plots
Recent update to change the way the LTM configuration dialog
worked accidentally lost the events setting.

Show events is not being stored in the LTMsettings 'datastream'
but as a separate chart property -- but *is* stored internally in
the LTMSettings class.

it was being lost when the datastream is appled. This fix makes
sure the value is remembered and reapplied after the settings are
applied.