Introduce SB Lowpoint
Let the user define the lowpoint for stress balance -- this is
so we can highlight or alert when they go below.

From a post-hoc perspective it is useful for identifying where
they went too low, but also for planning it helps to stop
planning to fail.

Will update the LTM chart to highlight it alongside SB metrics.