WKO+ Interval Markers
Rather than skip over the 'standard' interval markers because they
have such verbose names we now plot the vertical marker only.

This is to indicate there is an interval at a particular point to
signpost to the user that they can hover over this point to get an
interval summary.