FixElevation: support API key, fix file descriptor leak
Support a user defined MapQuest API key via src/Settings.h to
workaround the overloading of the existing key.

Use exceptions to trap Bad Request and API errors and stop processing
early.

Delete the networkMgr object to prevent a file descriptor leak
which can cause GC to crash when fix-elevation is configured
for auto-processing on import of files and there are many files
being imported.