Strava: move configuration into athlete-private.ini
When applying for API, the user must supply both the Client ID and
Client Secret in their API calls.  It was confusing to configure
Strava support one would add the CLIENT_SECRET in src/gcconfig.pri
and need to update the CLIENT_ID in src/Settings.h.

With the change, the Strava checkbox in the Share dialog is enabled
if the strava_client_secret is defined in athlete-private.ini.