Fix loading of "Ignore 0" setting in 3D plot
When opening gc and navigating to a tab with a custom 3d plot, the "Ignore 0" setting is not loaded correctly. This is due to bug that causes the preset values of the profile "User defined" to be applied after the "Ignore 0" setting has been loaded. To fix this, the currentIndexChanged signal of the combo box is connected after the settings have been loaded.