Fix calculation of minimum values of SmO2 and tHb and average of tHb
In file src/Metrics/BasicRideMetrics.cpp the minimum values of SmO2 and tHb and the average of tHb are calculated incorrectly. In functions compute of class MinSmO2 resp. class MintHb actually the maximum values get computed.
For MinSmO2 the comparison in function compute should read:

```c++
    if (point->smo2 >= 0.0f && (notset || point->smo2 < min)) {
        min = point->smo2;
        if (point->smo2 > 0.0f && notset)
            notset = false;
    }
```

with min being a double initialized to 0.0f and notset a boolean initialized to true.
For MinSmO2 we've got to handle that the first recorded values might be zero, e.g. when the athlete did not attach the Moxy sensor to the muscle prior to starting recording. If the comparison given above holds true we have to verify that we get a valid value > 0 first before allowing SmO2 to fall down to zero again.

For MintHb the comparison in function compute should read:

```c++
    if (point->thb > 0.0f && (notset || point->thb < min)) {
        min = point->thb;
        notset = false;
     }
```

with min being a double initialized to 0.0f and notset a boolean initialized to true.
Please note that for physiological reasons the minimum value of tHb must not be zero as this would imply that the muscle under observation were totally drained from blood.

At the end of functions compute in classes MinSmO2 and MintHb we have to set the correct value:

```c++
    setValue(min); // was max before
```

In class AvgtHb we have to take into account that tHb must not be zero as well for the same reason as stated above. In function compute the comparison should read:

```c++
    if (point->thb > 0.0f) {
        total += point->thb;
        ++count;
    }
```

References #2050
please review