Handle "Abort" and "Finish" correctly in RideImportWizard
The current wizard state was encoded in the text label of the abort
button, but the comparison operation did never match, because after
setText("Finish"), text() returns "&Finish". In Qt the ampersand encodes
accelerators in button text. It turns out that this is caused by KDE. It
automatically injects accelerators by default.

Only the default state was ever handled in abortClicked(). This prevents
aborts and causes the window to never close.

Add a variable to reliably store the current state, instead of encoding
it in the button text.