Fix creation of qwkcode from fractional seconds.
Qwkcode is limited to integer values of seconds. When a workout is specified in qwkcode then saved to an ERG file, the interval durations are rounded from integer seconds to fractional minutes with 2 decimal places of precision. When the ERG file is read, it is translated to a WWPoint object that stores time as double seconds. To accurately calculate the duration of a time interval defined by two WWPoints we should round the individual points first then take the integer difference. The table below demonstrates the consistency of rounding on both ends of the qwkcode -> ERG -> qwkcode round trip.

| qwkcode (original) | ERG time (rounded from qwkcode) | WWPoint time (precisely calculated from ERG time) | qwkcode (rounded from WWPoint) |
|--------------------|---------------------------------|---------------------------------------------------|--------------------------------|
|                  1 |                            0.02 |                                               1.2 |                              1 |
|                  2 |                            0.03 |                                               1.8 |                              2 |
|                  3 |                            0.05 |                                                 3 |                              3 |

(3 seconds is the smallest unit that fits evenly into a base ten fractional minute, so this pattern will repeat without loss of precision.)