Initial implementation of a new lap sound
This provides a fixed bell sound, played 3s before the
start of each new lap. This feature can be enabled under
the new training preferences dialog.

See the following bug report - quite exciting under Linux -
https://bugreports.qt.io/browse/QTBUG-40823

Due to this, is only currently enabled for QT versions 5.6.0
and above, could make this a platform specific check if felt this
is overly restrictive?