Support S&C sensors with single magnet
We see a few bug reports where people have combined speed
and cadence sensors, but only running one magnet (i.e. on
a direct drive trainer so no wheel magnet, or with a crank
based PM with no crank arm magnet). This interferes with
any valid speed and cadence data from other sensors (i.e
from a trainer or power meter).

This patch ignores the data from either the speed or cadence
side of the sensor, until at least one changed timestamp is
received on the channel for that data type. (The timestamps
never change if not triggered by the magnet).