Don't use special date handling for estimates that end in the future
The last estimate's time span (oftentimes) ends in the future. To
prevent adding CP ranges with future dates, the end date of this
estimate is amended to the current date. Previously the original
(future) date has been saved alongside the amended date, to uniquely
identify an estimate by its original date. This, however, is not true,
because if the last estimate is incomplete, it's an estimate by its own
and should not be identified by the future date (of another estimate
that may produce other values). Hence an incomplete estimate will now
be identified by the current date.