Fix EzUsb.c to allow Tacx firmware load on Windows
Previously, the call to usb_control_msg was suppressed for windows platforms in order to prevent
GoldenCheetah from failing when libusb was not available. This change allows usb_control_msg
to be explicitly linked at run time in the same way as other libusb functions within LibUsb.cpp

LibUsb.cpp will now resolve the usb_control_msg function in the same way it does other libusb references.

The 2 load functions ezusb_load_ram and ezusb_load_ram_imagic will now accept an extra argument.
For WIN32, the new argument will be a pointer to the usb_control_msg function in libusb0.
For other platforms, the new argument will be NULL.

For WIN32, EzUsb will now use the supplied pointer rather than defining its own reference
to usb_control_msg.

This should allow us to still start up OK when libusb is not available, but still be
able to load firmware when it is.